requirements.txt
langchain
python-dotenv
google-generativeai
langchain-google-genai
streamlit